+++ CHANGELOG.md
@@ -1,9 +1,22 @@
-# com.agava.merge2uiview Changelog
+# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-04-14
 
+### Added
+
+- Add `CooldownRepositoryFactory` for setting up cooldown of items.
+- Add `CooldownCommandFactory` to configure the cooldown command.
+- Add an abstract class `CooldownView` to display the cooldown.
+- Add fields to configure cooldown in `MergeRoot`.
+- Add an example of cooldown of items in `Samples`.
+
+### Changed
+
+- In the `ClickCommand` setting, you can set the list of `ClickCommandFactory` assets.
+- `com.yellowsquad.assetpath` updated to version `2.1.0`.
 
 ## [1.1.0] - 2023-04-11